+++ internal/provider/provider.go
@@ -137,6 +137,7 @@ func (p *provider) Resources(_ context.Context) []func() resource.Resource {
 		NewClusterResource,
 		NewPhysicalReplicationStreamResource,
 		NewSQLUserResource,
+		NewSQLRoleResource,
 		NewAllowlistResource,
 		NewPrivateEndpointServicesResource,
 		NewPrivateEndpointConnectionResource,